% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{chainCoda}
\alias{chainCoda}
\title{Export to Coda}
\usage{
chainCoda(s)
}
\arguments{
\item{s}{a list of chains generated by \code{heliosMetropolis}}
}
\value{
a \pkg{coda} object.
}
\description{
Convert to Coda objects
}
\details{
Convert samples generated by \code{helioMetropolis} or to a \pkg{coda}
object.
}
