% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{chainSummary}
\alias{chainSummary}
\alias{chainTail}
\alias{chainLast}
\alias{chainCollapse}
\alias{chainCov}
\alias{chainBcov}
\alias{chainAcceptance}
\title{Manipulate MCMC samples}
\usage{
chainSummary(s)

chainTail(s, discard = 0, thin = 1)

chainLast(s)

chainCollapse(s, collapse = TRUE, discard = 0, thin = 1, chains = NULL)

chainCov(s, discard = 0, chains = NULL)

chainBcov(s, discard = 0, chains = NULL)

chainAcceptance(s, collapse = FALSE, chains = NULL)
}
\arguments{
\item{s}{a single chain or a list of parallel chains generated by
\code{heliosMetropolis}.}

\item{discard}{number of initial samples to discard.}

\item{thin}{rate at which to thin the sample.}

\item{collapse}{whether to collapse parallel chains to a single
chain}

\item{chains}{the set of chains to retain, or \code{NULL}.}
}
\value{
\itemize{
\item \code{chainSummary} returns a summary of the sample
\item \code{chainTail} discards the initial samples from each chain
\item \code{chainLast} returns the last sample for each location in each chain
\item \code{chainCollapse} collapses multiple chains into a single sample
\item \code{chainCov} returns the covariance of the parameters location by location
as a pxpxn array.
\item \code{chainBcov} returns the joint covariance of the parameters
as an (np)x(np) array.
\item \code{chainAcceptance} returns the acceptance rate in the (thinned) chain
}
}
\description{
Manipulate the samples generated by the Metropolis samplers.
}
\details{
These functions provide some basic operations on the samples
generated by the Metropolis samplers for Helios.
}
