% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{heliosDiscGridWeights}
\alias{heliosDiscGridWeights}
\alias{heliosDiscRasterWeights}
\alias{extractHeliosDiscRasterWeights}
\title{Helios Discrete Grid Weights}
\usage{
heliosDiscGridWeights(grid, model, autoscale = NA, points = TRUE)

heliosDiscRasterWeights(raster, model, autoscale = NA, points = FALSE)

extractHeliosDiscRasterWeights(raster, weights, k, log = TRUE)
}
\arguments{
\item{grid}{a two column matrix of (lon,lat) grid points.}

\item{model}{a model structure as generated by \code{\link[=heliosModel]{heliosModel()}}
or equivalent.}

\item{autoscale}{a lower bound for the (finite) weights}

\item{points}{should the contribution to the posterior from the
track points be included.}

\item{raster}{a terra raster used to define the grid.}

\item{weights}{a weights object generated with
\code{heliosDiscGridWeights} or \code{heliosDiscRasterWeights}}

\item{k}{the rack point ot extract}

\item{log}{show the weights be returned on the log scale.}
}
\value{
The \code{heliosDiscGridWeights} and \code{heliosDiscRasterWeights}
functions return a list with the grid of locations, the model
object and the matrix of weights. The
\code{extractHeliosDiscRasterWeights} function returns a column of
weights as a raster.
}
\description{
Construct the Helios weights on a discrete spatial grid for use
with 'heliosDiscMinimalTrack' and \code{heliosDiscSampler}.
}
\details{
The \code{\link[=heliosDiscMinimalTrack]{heliosDiscMinimalTrack()}} and \code{\link[=heliosDiscSampler]{heliosDiscSampler()}} functions
generate approximate tracks on a discrete spatial grid that can be
used to initialize \code{\link[=heliosMetropolis]{heliosMetropolis()}} sampler.  Both these
functions require a weights object that defines the spatial grid
and the relative likelihood of the points on this grid based on
the observed light data.

To generate a weights object with \code{heliosDiscGridWeights} the user
must specify the grid from which locations will be drawn as a two
column matrix of (lon,lat) points. The grid can be arbitrary but
must be cover the range of the track.

The \code{heliosDiscRasterWeights} function provides an alternative
interface in which the grid is specified as a terra raster.  If
that raster has values, locations with values that not finite are
excluded from the grid.  The \code{extractHeliosDiscRasterWeights}
function extracts the weights for a single track point as a
raster, and can be used to visualize the computed weights.

The weights are stored as a matrix with a row for each grid point
and column for each point along the track, and the entries of each
column are the relative posterior probabilities that that track
point passes through the corresponding grid point.

If \code{points=TRUE} the weights are calculated from the light data
along tracks segments and the contribution to the posterior from
the track points.  If \code{points=FALSE} the weights are calculated
from the contribution to the posterior from the light data alone.
If the contribution to the posterior from the tracks points is
only used to exclude locations, it is more efficient to exclude
those points from the grid directly and set \code{points=FALSE}.

If \code{autoscale} is \code{NA} the scaling of the weights is determined
directly by the posterior. Otherwise the weights are scaled so the
maximum weight is 1 and the minimum weight of a feasible grid
point is the value of \code{autoscale}.  If scaling is a vector, its
elements define the scaling used at each individual location.
}
