% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{heliosDiscMinimalTrack}
\alias{heliosDiscMinimalTrack}
\title{Minimal distance paths on a discrete grid.}
\usage{
heliosDiscMinimalTrack(weights, gamma, ngbrs = 50)
}
\arguments{
\item{weights}{a weights object generated with
\code{heliosDiscGridWeights} or \code{heliosDiscRasterWeights}}

\item{gamma}{regularization parameter controlling the distance cost.}

\item{ngbrs}{the number of neighbours of each grid point to consider.}
}
\value{
track locations as a two column matrix.
}
\description{
Compute an approximate track on a discrete spatial grid by
maximizing the Helios posterior regularizing by distance
travelled.
}
\details{
The \code{heliosDiscMinimalTrack} function seeks the shortest track
defined on a discrete spatial grid that is consistent with the
observed light, by maximizing the contribution to the posterior
from the light data regularized by total distance travelled.

This function uses dynamic programming to maximise the log Helios
posterior penalized by the total distance travelled, when the
track points are constrained to a discrete spatial grid.  This
function requires a weights object calculated with
\code{\link[=heliosDiscGridWeights]{heliosDiscGridWeights()}} or \code{\link[=heliosDiscRasterWeights]{heliosDiscRasterWeights()}} with
\code{log=TRUE}.

The \code{gamma} argument scales the distance penalty used to
regularize the posterior.  Larger values of \code{gamma} result in
shorter tracks, at the expense of the accuracy with which the
fitted track reproduces the observed light.  For efficiency, the
distance penalty is calculated for only the \code{ngbrs} nearest
neighbours of each grid point.

The resulting track is discretized to the grid, and so for coarser
grids it may be desirable to set the fixed points back to their
exact locations for use in further analysis.
}
