% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{heliosLightGrid}
\alias{heliosLightGrid}
\alias{heliosLightRaster}
\title{Helios Light Grid}
\usage{
heliosLightGrid(grid, date, light, alpha, zenith, forbid = -Inf)

heliosLightRaster(raster, date, light, alpha, zenith, forbid = -Inf)
}
\arguments{
\item{grid}{a two column matrix of (lon,lat) grid points.}

\item{date}{POSIXct vector of times for day/nights observations}

\item{light}{logical vector of day/night observations}

\item{alpha}{rate parameter for sensor obscuration.}

\item{zenith}{the solar zenith angle that defines day/night.}

\item{forbid}{the log likelihood for forbidden light observations}

\item{raster}{a terra raster.}
}
\value{
A vector or raster of likelihood values.
}
\description{
Calculate the Helios log likelihood over a discrete spatial grid
}
\details{
Calculates contribution to the Helios log likelhood from a light
record across a grid of locations specified by user.

The \code{heliosLightGrid} function requires the grid to be specied as
a two column matrix of (lon,lat) locations, with the result
returned as a vector.

The \code{heliosLightRaster} function provides an alternative interface
in which the grid is specified as a terra raster, with the result
returned as raster.  If the input raster has values, locations
with values that are missing or not finite are excluded from the
grid.
}
