% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{heliosModel}
\alias{heliosModel}
\alias{heliosLightModel}
\title{Helios Model Structures}
\usage{
heliosModel(
  date,
  light,
  time,
  x0,
  alpha,
  beta,
  logp.p0 = function(x) rep.int(0L, nrow(x)),
  logp.s0 = function(x) rep.int(0L, nrow(x) - 1),
  fixedx = FALSE,
  dt = NULL,
  zenith = 96,
  forbid = -Inf
)

heliosLightModel(date, light, time, alpha, zenith, forbid)
}
\arguments{
\item{date}{POSIXct vector of times for day/night observations}

\item{light}{logical vector of day/night observations}

\item{time}{times at which to fit locations}

\item{x0}{initial estimates of fitted locations.}

\item{alpha}{rate parameter for sensor obscuration.}

\item{beta}{parameters of the behavioural model.}

\item{logp.p0}{function to evaluate any additional contribution to
the log posterior from the estimated locations on the track}

\item{logp.s0}{function to evaluate any additional contribution to
the log posterior from the estimated segments of track}

\item{fixedx}{logical vector indicating which estimated locations
to hold fixed.}

\item{dt}{(optional) vector of time intervals (hours) for speed
calculation.}

\item{zenith}{the solar zenith angle that defines day/night.}

\item{forbid}{the log likelihood for forbidden light observations}
}
\value{
The \code{heliosModel} function returns list with components
\item{\code{time}}{the times at which locations are estimated}
\item{\code{x0}}{an array of initial location estimates.}
\item{\code{fixedx}}{a logical vector indicating which locations are
fixed.}
\item{\code{logp.s}}{function to evaluate the contributions to the log
posterior from the track segments}
\item{\code{logp.p}}{function to evaluate the contributions to the log
posterior from the track points}
\item{\code{logp.l}}{function to evaluate the contributions to the log
posterior from the light model on each segment}
\item{\code{logp.b}}{function to evaluate the contributions to the log
posterior from the behavioural model on each segment}
\item{\code{predict}}{function to predict the zenith angle along a given
track.}
\item{\code{Date}}{POSIXct vector of times for day/night observations.}
\item{\code{Light}}{logical vector of day/night observations.}
The \code{heliosLightModel} function returns just the \code{predict} and
\code{logp.l} components.
}
\description{
Helios Model Structures.
}
\details{
Helios requires a model structure that describes the model being
fitted by providing functions that compute the contributions to
the log posterior from each of the segments along the track
(\code{logp.s}) and each of the locations on the track (\code{logp.p}),
together with a vector of \code{time} times for which locations are
estimated, a two column matrix \code{x0} of initial estimates of those
locations and a logical vector \code{fixedx} that indicates which of
those locations are fixed. This function generate basic model
structures that should provide a suitable starting point for most
analyses.

The \code{heliosModel} function constructs a model structure that
combines the \code{heliosLightModel} with a behavioural model that
assume the average speed of travel between successive locations is
Gamma distributed.

The helios light model assumes the light sensor cannot detect
light when it is dark, but may fail to detect light when it is
light because the sensor is obscured.  The model assumes the
length of time the sensor can be obscured is exponentially
distributed, with rate \code{alpha}.

The \code{forbid} argument represents the likelihood of light being
observed when it is dark.  When \code{forbid=-Inf} the initial
locations \code{x0} must be consistent with the observed light - there
can be no light observed when it is dark on the implied track.
Setting \code{forbid} to a large negative value relaxes this
constraint, making the observation of light in the dark unlikely
but not impossible.

The behavioural model assumes the speed of travel is Gamma
distributed.  The shape and rate of the Gamma distribution of
speeds is specified by the \code{beta} parameter. If \code{beta} is a two
element vector the same shape and rate are applied across all
segments, but if \code{beta} is a two column matrix the shape and rate
can be specified on a segment by segment basis. By default, the
speed of travel is calculated based on the time intervals between
the estimated locations (in hours), but the intervals of time
actually available for travel can be specified directly with the
\code{dt} argument.  The behavioural model makes a contribution to the
posterior on each segment of track.

Both the light model and the behavioural model and make a
contribution to the posterior on each segment of track.  The user
is free to provide a function \code{logp.s0} of a single argument \code{x}
(the locations) that returns an additional contribution to the
posterior for each track segment, and a funtion \code{logp.p0} that
returns an additional contribution to the posterior for each
location.
}
