% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{locationSummary}
\alias{locationSummary}
\alias{locationMean}
\title{Summaries of Location Samples}
\usage{
locationSummary(
  obj,
  time = NULL,
  discard = 0L,
  alpha = 0.95,
  collapse = TRUE,
  chains = NULL
)

locationMean(obj, discard = 0L, collapse = TRUE, chains = NULL)
}
\arguments{
\item{obj}{the object, or a single chain or a list of parallel
chains generated by \code{heliosMetropolis}.}

\item{time}{the times corresponding to the x locations.}

\item{discard}{number of initial samples to discard.}

\item{alpha}{coverage of the credible intervals calculated by
\code{locationSummary}.}

\item{collapse}{whether to collapse parallel chains to a single
chain}

\item{chains}{the set of chains to retain, or \code{NULL}.}
}
\value{
\item{\code{locationSummary}}{returns a dataframe or a list of
dataframes of summary quantities for each location.}
\item{\code{locationMean}}{returns an array or a list of arrays
of the means of the samples for each location.}
}
\description{
Summarize a set of location samples
}
\details{
These functions compute various summaries of a sample or list of
samples generated by \code{heliosMetropolis}.

These functions accept either a sample from a single mcmc run, or
a list of samples from parallel mcmc runs.  When \code{collapse} is
true, multiple samples are merged and single result is returned,
otherwise a result is returned for each sample.
}
