% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SGAT-package}
\alias{SGAT-package}
\title{Solar Geolocation for Animal Tracking}
\description{
Provides facilities for estimating broadscale animal motions from
local light level data recorded by an archival tag.
}
\author{
S. Wotherspoon, M. Sumner, A. Bindoff and S. Lisovski.
}
