% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII-package.R
\docType{package}
\name{SGATII-package}
\alias{SGATII}
\alias{SGATII-package}
\title{SGATII: Solar/Satellite Geolocation for Animal Tracking II}
\description{
Facilities for estimating the broadscale animal motions from archival tags that record local light levels.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWotherspoon/SGATII}
  \item Report bugs at \url{https://github.com/SWotherspoon/SGATII/issues}
}

}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

Authors:
\itemize{
  \item Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})
  \item Michael Sumner \email{Michael.Sumner@aad.gov.au}
  \item Aidan Bindoff \email{Aidan.Bindoff@utas.edu.au} (\href{https://orcid.org/0000-0002-0943-2702}{ORCID})
  \item Ben Raymond \email{Ben.Raymond@aad.gov.au}
  \item Simeon Lisovski \email{Simeon.Lisovski@awi.de} (\href{https://orcid.org/0000-0002-6399-0035}{ORCID})
}

}
\keyword{internal}
