% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{trackBearingChange}
\alias{trackBearingChange}
\title{Distance along a path}
\usage{
trackBearingChange(x)
}
\arguments{
\item{x}{a two column matrix of (lon,lat) locations along the path.}
}
\value{
vector of changes in bearing (degrees)
}
\description{
Bearing changes along a track
}
\details{
Compute the change in bearing between successive locations along
path.
}
