% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{trackDist}
\alias{trackDist}
\title{Distance along a path}
\usage{
trackDist(x)
}
\arguments{
\item{x}{a two column matrix of (lon,lat) locations along the path.}
}
\value{
vector of interpoint distances (km)
}
\description{
Distances along a path
}
\details{
Compute the great circle distances (in km) between successive
locations along path.
}
