% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{heliosDiscSampler}
\alias{heliosDiscSampler}
\title{Discrete Sampler}
\usage{
heliosDiscSampler(weights, x = NULL, iters = 100L)
}
\arguments{
\item{weights}{a weights object generated with
\code{heliosDiscGridWeights} or \code{heliosDiscRasterWeights}}

\item{x}{starting values for estimated locations x.}

\item{iters}{number of samples to draw.}
}
\value{
Returns an n x 2 x r array of r samples for the n
locations.
}
\description{
Draw samples for Helios from a discrete spatial grid.
}
\details{
The \code{heliosDiscSampler} function samples from a Helios model where
the locations are constrained to a discrete spatial grid, and can
be used to generate an initial track for the
\code{\link[=heliosMetropolis]{heliosMetropolis()}} sampler.  The sampler requires a weights
object generated with either \code{\link[=heliosDiscGridWeights]{heliosDiscGridWeights()}} or
\code{\link[=heliosDiscRasterWeights]{heliosDiscRasterWeights()}} with \code{log=FALSE}.

The weights define the proposal distribution used by the sampler
and represent the relative probabilities of a grid point being
selected by the proposal distribution at location along the track.
The weights are derived from the model likelihood assuming the tag
is stationary.

This sampler does not require initial location estimates and it
only uses the initial locations \code{x0} from the model to specify
fixed points, the remaining elements of \code{x0} are ignored.

This is an independence sampler in which the proposal distribution
is based on an approximation the to helios likelihood obtained by
assuming the tag is stationary. So for very fast moving
individuals, the method can perform poorly.
}
