% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{heliosDiscSampler}
\alias{heliosDiscSampler}
\alias{heliosDiscGridWeights}
\alias{heliosDiscRasterWeights}
\title{Discrete Sampler}
\usage{
heliosDiscSampler(weights, x = NULL, iters = 100L)

heliosDiscGridWeights(grid, model)

heliosDiscRasterWeights(raster, model)
}
\arguments{
\item{weights}{a weights object generated with
\code{heliosDiscGridWeights} or \code{heliosDiscRasterWeights}}

\item{x}{starting values for estimated locations x.}

\item{iters}{number of samples to draw.}

\item{grid}{a two column matrix of (lon,lat) grid points.}

\item{model}{a model structure as generated by \code{\link[=heliosModel]{heliosModel()}}
or equivalent.}

\item{raster}{a terra raster.}
}
\value{
Returns an n x 2 x r array of r samples for n locations.
}
\description{
Draw samples for Helios from a discrete spatial grid.
}
\details{
The \code{heliosDiscSampler} function draws samples from a Helios model
where the locations are constrained to a discrete grid, and can be
used to generate an initial track for the \code{\link[=heliosMetropolis]{heliosMetropolis()}}
sampler. This sampler requires a weights object generated with
either \code{heliosDiscGridWeights} or \code{heliosDiscRasterWeights}.

To generate weights with \code{heliosDiscGridWeights} the user must
specify the grid from which locations will be drawn as a two
column matrix of (lon,lat) points. The grid can be arbitrary, but
if the grid is too fine the sampler can be very memory intensive,
and if the grid is to coarse the track can be inaccurate.  Even
fro moderate grids the weights object can be very large, and
should be discarded when no longer required.

The \code{heliosDiscRasterWeights} function provides an alternative
interface in which the grid is specified as a terra raster.  If
that raster has values, locations with values that not finite are
excluded from the grid.

This sampler does not require initial location estimates, and it
only uses the initial locations \code{x0} from the model to specify
fixed points.

The sampler is an independence sampler in which the proposal is
based on an approximation the to helios likelihood obtained by
assuming the tag is stationary.  So for very fast moving
individuals, the method can perform poorly.
}
