% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGATII.R
\name{heliosMetropolis}
\alias{heliosMetropolis}
\title{Metropolis Samplers}
\usage{
heliosMetropolis(
  model,
  proposal,
  x0 = NULL,
  iters = 1000L,
  thin = 10L,
  chains = 1L,
  parallel = TRUE
)
}
\arguments{
\item{model}{a model structure as generated by \code{\link[=heliosModel]{heliosModel()}}
or equivalent.}

\item{proposal}{function for drawing proposals for x.}

\item{x0}{Starting values for estimated locations x.}

\item{iters}{number of samples to draw.}

\item{thin}{rate at which to thin samples.}

\item{chains}{number of chains to sample.}

\item{parallel}{enable parallel evaluation through futures.}
}
\value{
If there are r samples drawn for each of q chains of p
parameters at n locations, Helios will return a list containing
\item{\code{model}}{the model structure}
\item{\code{x}}{a list of n x p x r arrays of estimated locations from
the q chains}
}
\description{
Metropolis sampler for Helios
}
\details{
These functions draw samples from the posterior for the Helios
model by the Metropolis algorithm.  If \code{parallel} is \code{TRUE} the
chains are drawn in parallel with \code{\link[=future_lapply]{future_lapply()}}.

The starting values \code{x0} should be specified as a two column
matrix, or as a list of two column matrices, one for each chain.
}
